+++ data/src/main/java/com/vljx/hawkspeed/data/network/models/vehicle/stock/VehicleStockDto.kt
@@ -34,7 +34,7 @@ data class VehicleStockDto(
 
     @Expose
     @SerializedName("displacement")
-    val displacement: Float,
+    val displacement: Int,
 
     @Expose
     @SerializedName("induction")
